+++ workflows_observability/activities.py
@@ -1,8 +1,7 @@
-from atlan_helpers import get_atlan_client
+from helpers import get_atlan_client, save_result_locally
 from application_sdk.activities import ActivitiesInterface
 from application_sdk.observability.logger_adaptor import get_logger
 from temporalio import activity
-from pyatlan.model.enums import AtlanWorkflowPhase
 import os
 from datetime import datetime
 
@@ -43,6 +42,7 @@ class WorkflowsObservabilityActivities(ActivitiesInterface):
             difference_in_hours = int(time_difference.total_seconds() / 3600)
 
             client = get_atlan_client()
+            from pyatlan.model.enums import AtlanWorkflowPhase
 
             results = client.workflow.find_runs_by_status_and_time_range(
                 status=[AtlanWorkflowPhase.SUCCESS, AtlanWorkflowPhase.FAILED],
@@ -55,35 +55,4 @@ class WorkflowsObservabilityActivities(ActivitiesInterface):
 
         except Exception as e:
             logger.error(f"Failed to process workflows: {str(e)}", exc_info=e)
-            raise
-
-
-def save_result_locally(result, local_directory: str) -> None:
-    """
-    Save a workflow run result to a local directory, structured by date and status.
-
-    Args:
-        result: The workflow result object returned by the Atlan client.
-        local_directory (str): Base path where files will be saved.
-
-    Raises:
-        OSError: If directories or file writing fails.
-    """
-    try:
-        date_str = result.source.status.startedAt[:10]
-        subdirs = [date_str, f"{date_str}/SUCCESS", f"{date_str}/FAILED"]
-
-        for subdir in subdirs:
-            os.makedirs(os.path.join(local_directory, subdir), exist_ok=True)
-
-        status_dir = "SUCCESS" if result.status == AtlanWorkflowPhase.SUCCESS else "FAILED"
-        output_path = os.path.join(local_directory, date_str, status_dir, result.id + ".json")
-
-        with open(output_path, "w") as f:
-            f.write(result.json(by_alias=True, exclude_none=True))
-
-        logger.info(f"Saved result to {output_path}")
-
-    except Exception as e:
-        logger.error(f"Error saving workflow result locally: {str(e)}", exc_info=e)
-        raise
+            raise e
