--- workflows_observability/atlan_helpers.py
@@ -1,8 +1,0 @@
-from pyatlan.client.atlan import AtlanClient
-import os
-
-def get_atlan_client() -> AtlanClient:
-    return AtlanClient(
-        base_url=os.getenv("ATLAN_BASE_URL"),
-        api_key=os.getenv("ATLAN_API_KEY") 
-    )
